move UUFRuntimeException to api.exception package
